+++ alphafold/partition.py
@@ -3,6 +3,7 @@ from copy import deepcopy
 from alphafold.secstruct import *
 from alphafold.explicit_recursions import *
 from alphafold.dynamic_programming import *
+from alphafold.backtrack import mfe
 
 ##################################################################################################
 class AlphaFoldParams:
@@ -157,52 +158,6 @@ def initialize_sequence_information( self ):
 
     self.any_intervening_cutpoint = initialize_any_intervening_cutpoint( self.is_cutpoint )
 
-###################################################################################################################33
-class DynamicProgrammingData:
-    '''
-    Dynamic programming object, with derivs and contribution accumulation.
-     X   = values (N x N)
-     dQ  = derivatives (N X N)
-     X_contrib = contributions (coming soon)
-    '''
-    def __init__( self, N ):
-        self.Q = []
-        for i in range( N ): self.Q.append( [0.0]*N )
-
-        self.dQ = deepcopy( self.Q ) # another zero matrix.
-
-        self.contrib = []
-        for i in range( N ): self.contrib.append( [[]]*N )
-
-    def __getitem__( self, idx ):
-        # overloaded []. warning: overhead! directly access object.Q[ idx ] in inner loops.
-        return self.Q[ idx ]
-
-    def __len__( self ): return len( self.Q )
-
-    def add( self, i, j, b ):
-        #  trying out a function that might make code more readable,
-        #  (could hide all contribution accumulation for backtracking -- and
-        #   perhaps even derivatives -- inside class!)
-        # but this kind of thing appears to take up too much overhead.
-        self.Q[i][j]  += b
-        self.dQ[i][j] += 0
-        self.Q_contrib[i][j].append( [i,j,b] )
-
-##################################################################################################
-class BasePairType:
-    def __init__( self, nt1, nt2, Kd_BP, N ):
-        '''
-        Uh, a little weird to have Z in here.
-        '''
-        self.nt1 = nt1
-        self.nt2 = nt2
-        self.Kd_BP = Kd_BP
-        self.Z_BP_DP  = DynamicProgrammingData( N );
-        self.Z_BP  = self.Z_BP_DP.Q
-        self.dZ_BP = self.Z_BP_DP.dQ;
-        self.match_lowercase = ( nt1 == '' and nt2 == '' )
-
 ##################################################################################################
 def initialize_dynamic_programming_matrices( self ):
     '''
@@ -224,6 +179,20 @@ def initialize_dynamic_programming_matrices( self ):
         self.C_eff.Q[i][i]                 = self.params.C_init
     self.C_eff_no_coax_singlet = deepcopy( self.C_eff )
     self.C_eff_no_BP_singlet   = deepcopy( self.C_eff )
+
+##################################################################################################
+class BasePairType:
+    def __init__( self, nt1, nt2, Kd_BP, N ):
+        '''
+        Uh, a little weird to have Z in here.
+        '''
+        self.nt1 = nt1
+        self.nt2 = nt2
+        self.Kd_BP = Kd_BP
+        self.Z_BP_DP  = DynamicProgrammingData( N );
+        self.Z_BP  = self.Z_BP_DP.Q
+        self.dZ_BP = self.Z_BP_DP.dQ;
+        self.match_lowercase = ( nt1 == '' and nt2 == '' )
 
 ##################################################################################################
 def initialize_base_pair_types( self ):
@@ -248,71 +217,6 @@ def initialize_any_intervening_cutpoint( is_cutpoint ):
             any_intervening_cutpoint[ i ][ j ] = found_cutpoint
             if is_cutpoint[ j ]: found_cutpoint = True
     return any_intervening_cutpoint
-
-##################################################################################################
-##################################################################################################
-##################################################################################################
-def get_random_contrib( contribs ):
-    # Random sample weighted by probability. Must be a simple function for this.
-    contrib_cumsum = [ contribs[0][0] ]
-    for contrib in contribs[1:]: contrib_cumsum.append( contrib_cumsum[-1] + contrib[0] )
-    r = random.random() * contrib_cumsum[ -1 ]
-    for (idx,psum) in enumerate( contrib_cumsum ):
-        if r < psum: return contribs[idx]
-
-##################################################################################################
-def backtrack( self, contribs_input, mode = 'mfe' ):
-    if len( contribs_input ) == 0: return []
-    print 'contribs_input', contribs_input
-    contrib_sum = sum( contrib[0] for contrib in contribs_input )
-    if   mode == 'enumerative': contribs = deepcopy( contribs_input )
-    elif mode == 'mfe':         contribs = [ max( contribs_input ) ]
-    elif mode == 'stochastic' : contribs = [ get_random_contrib( contribs_input ) ]
-    p_bps = [] # list of tuples of (p_structure, bps_structure) for each structure
-    N = self.N
-    for contrib in contribs: # each option ('contribution' to this partition function of this sub-region)
-        if ( contrib[0] == 0.0 ): continue
-        p_contrib = contrib[0]/contrib_sum
-        p_bps_contrib = [ [p_contrib,[]] ]
-
-        for backtrack_info in contrib[1]: # each 'branch'
-            ( Z_backtrack_id, i, j )  = backtrack_info
-            if Z_backtrack_id == id(self.Z_BP):
-                update_Z_BP( self, i, j, calc_contrib = True )
-                backtrack_contrib = self.Z_BP.contrib
-                p_bps_contrib = [ [p_bp[0], p_bp[1]+[(i%N,j%N)] ] for p_bp in p_bps_contrib ]
-            elif Z_backtrack_id == id(self.C_eff):
-                update_C_eff( self, i, j, calc_contrib = True )
-                backtrack_contrib = self.C_eff.contrib
-            elif Z_backtrack_id == id(self.C_eff_no_coax_singlet):
-                update_C_eff( self, i, j, calc_contrib = True )
-                backtrack_contrib = self.C_eff_no_coax_singlet.contrib
-            elif Z_backtrack_id == id(self.Z_linear):
-                update_Z_linear( self, i, j, calc_contrib = True )
-                backtrack_contrib = self.Z_linear.contrib
-            p_bps_component = backtrack( self, backtrack_contrib[i%N][j%N], mode )
-            if len( p_bps_component ) == 0: continue
-            # put together all branches
-            p_bps_contrib_new = []
-            for p_bps1 in p_bps_contrib:
-                for p_bps2 in p_bps_component:
-                    p_bps_contrib_new.append( [p_bps1[0]*p_bps2[0], p_bps1[1]+p_bps2[1]] )
-            p_bps_contrib = p_bps_contrib_new
-
-        p_bps += p_bps_contrib
-    return p_bps
-
-##################################################################################################
-def mfe( self, Z_final_contrib ):
-    p_bps = backtrack( self, Z_final_contrib, mode = 'mfe' )
-    assert( len(p_bps) == 1 )
-    return (p_bps[0][1],p_bps[0][0])
-
-##################################################################################################
-def boltzmann_sample( self, Z_final_contrib ):
-    p_bps = backtrack( self, Z_final_contrib, mode = 'stochastic' )
-    assert( len(p_bps) == 1 )
-    return (p_bps[0][1],p_bps[0][0])
 
 ##################################################################################################
 def _calc_mfe( self ):
